Moved all common steps to a shared file so it can be used in multiple SIA tests - coderabbit
